Add files via upload
The Netflix Project is an impressive website that replicates the popular streaming platform using HTML, CSS, and JavaScript. This project aims to provide users with a visually appealing and interactive experience similar to that of the actual Netflix platform.

By utilizing HTML, the website's structure is built, ensuring proper organization of content and elements. CSS is applied to enhance the design, including styling for fonts, colors, layouts, and responsive design techniques. This ensures that the website is visually appealing and compatible across different screen sizes.

JavaScript is used to add functionality and interactivity to the Netflix project. This includes features such as dynamic sliders or carousels to showcase different movies or TV shows, interactive buttons for play/pause functionality, and search functionality to allow users to find their favorite content easily.

The Netflix project meticulously replicates the familiar Netflix user interface, providing users with a seamless and intuitive browsing experience. Users can navigate through different genres, view trailers, read descriptions, and select their preferred movies or TV shows for playback.

With the combination of HTML, CSS, and JavaScript, this project successfully recreates the essence of Netflix and offers users a taste of the streaming platform's interface. It showcases the ability to create engaging and interactive websites using these technologies, providing an immersive experience for users who enjoy streaming movies and TV shows.